@@ -1,56 +1,61 @@
 {
   "name": "pc-datacatalog",
-  "version": "2022.1.1",
+  "version": "2022.2.0",
   "private": true,
-  "proxy": "http://localhost:7071/",
+  "proxy": "http://api:7071/",
   "dependencies": {
     "@apidevtools/json-schema-ref-parser": "^9.0.9",
-    "@craco/craco": "^6.4.0",
-    "@fluentui/react": "^8.41.0",
+    "@craco/craco": "^6.4.3",
+    "@fluentui/react": "^8.61.0",
     "@radiantearth/stac-fields": "1.0.0-beta.7",
-    "@reduxjs/toolkit": "^1.6.1",
-    "@testing-library/jest-dom": "^5.15.0",
-    "@testing-library/react": "^12.1.2",
+    "@reduxjs/toolkit": "^1.8.0",
+    "@testing-library/jest-dom": "^5.16.2",
+    "@testing-library/react": "^12.1.3",
     "@testing-library/user-event": "^13.5.0",
     "@turf/bbox-polygon": "^6.5.0",
     "@turf/boolean-intersects": "^6.5.0",
     "@turf/centroid": "^6.5.0",
     "@turf/union": "^6.5.0",
+    "@types/dompurify": "^2.3.3",
     "@types/geojson": "^7946.0.8",
-    "@types/jest": "^27.0.2",
-    "@types/lodash-es": "^4.17.5",
-    "@types/marked": "^3.0.2",
-    "@types/node": "^16.11.6",
-    "@types/react": "^17.0.34",
-    "@types/react-dom": "^17.0.11",
-    "@types/react-redux": "^7.1.20",
-    "@types/react-router-dom": "^5.3.2",
+    "@types/jest": "^27.4.1",
+    "@types/lodash-es": "^4.17.6",
+    "@types/marked": "^4.0.2",
+    "@types/node": "^17.0.21",
+    "@types/react": "^17.0.39",
+    "@types/react-dom": "^17.0.12",
+    "@types/react-redux": "^7.1.23",
+    "@types/react-router-dom": "^5.3.3",
     "@uifabric/icons": "^7.6.2",
-    "axios": "^0.24.0",
-    "azure-maps-control": "^2.1.6",
-    "dayjs": "^1.10.7",
-    "dompurify": "^2.3.3",
+    "axios": "^0.26.0",
+    "azure-maps-control": "^2.1.10",
+    "dayjs": "^1.10.8",
+    "dompurify": "^2.3.6",
     "formik": "^2.2.9",
-    "highlight.js": "^11.3.1",
+    "highlight.js": "^11.4.0",
     "json-stringify-pretty-compact": "^3.0.0",
     "lodash-es": "^4.17.21",
-    "marked": "^3.0.4",
+    "marked": "^4.0.12",
     "number-abbreviate": "^2.0.0",
-    "query-string": "^7.0.1",
+    "query-string": "^7.1.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-error-boundary": "^3.1.4",
     "react-helmet": "^6.1.0",
-    "react-query": "^3.31.0",
+    "react-query": "^3.34.16",
     "react-redux": "^7.2.6",
-    "react-router-dom": "^5.3.0",
+    "react-router-dom": "^6.2.2",
     "react-router-hash-link": "^2.4.3",
     "react-scripts": "4.0.3",
-    "react-use": "^17.3.1",
+    "react-use": "^17.3.2",
     "redux": "^4.1.2",
-    "swagger-ui-react": "^4.0.1",
-    "typescript": "^4.4.4",
+    "swagger-ui-react": "^4.6.0",
+    "typescript": "^4.6.2",
     "yup": "^0.32.11"
+  },
+  "resolutions": {
+    "//": "See https://github.com/facebook/create-react-app/issues/11773",
+    "react-error-overlay": "6.0.9"
   },
   "scripts": {
     "start": "craco start",
@@ -87,9 +92,11 @@
     ]
   },
   "devDependencies": {
+    "@azure/static-web-apps-cli": "^0.8.2",
     "copy-webpack-plugin": "^6.4.1",
-    "cypress": "^8.5.0",
+    "cypress": "^9.5.4",
     "html-loader": "^1.3.2",
+    "http-proxy-cli": "2.1.0",
     "js-yaml": "^4.1.0",
     "json-loader": "^0.5.7",
     "markdown-loader": "^6.0.0",
