@@ -1,10 +1,11 @@
 import { Separator } from "@fluentui/react";
-import React from "react";
-import { Redirect, useParams } from "react-router-dom";
+import { Navigate, useParams } from "react-router-dom";
 import GeneratedInternalToc from "./GeneratedInternalToc";
 import TopicNav from "./TopicNav";
 
 import DocsHtmlContent from "./DocsHtmlContent";
+import { useEffect } from "react";
+import { scrollToHash } from "utils";
 
 const centerPanelWidth = "75%";
 
@@ -12,6 +13,10 @@ const Topic = ({ topics }) => {
   const { topicId, fileId } = useParams();
   const docsKey = `./${topicId}/${fileId}.json`;
   const doc = topics[docsKey];
+
+  useEffect(() => {
+    scrollToHash(window.location.hash);
+  });
 
   const bottomNav = <TopicNav topic={topicId} prev={doc.prev} next={doc.next} />;
 
@@ -31,7 +36,7 @@ const Topic = ({ topics }) => {
       </>
     );
   }
-  return <Redirect to={"/404"} />;
+  return <Navigate replace to={"/404"} />;
 };
 
 export default Topic;
