@@ -1,12 +1,12 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import { useHistory } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 
 import AssetThumbnail from "./AssetThumbnail";
 import Keywords from "./Keywords";
 
 const CollectionCard = ({ collection }) => {
-  const history = useHistory();
+  const navigate = useNavigate();
   const href = `/dataset/${collection.id}`;
 
   return (
@@ -14,7 +14,7 @@ const CollectionCard = ({ collection }) => {
       <Link to={href} aria-label={`Navigate to ${collection.id}`}>
         <AssetThumbnail assets={collection.assets} />
 
-        <h3 style={{ marginBottom: 0 }}>{collection.title}</h3>
+        <h3 style={{ marginBottom: 0 }}>{collection.title || collection.id}</h3>
       </Link>
       <p style={{ color: "initial", minHeight: 40 }}>
         {collection["msft:short_description"]}
@@ -23,7 +23,7 @@ const CollectionCard = ({ collection }) => {
         keywords={collection.keywords}
         color="#4C4C51"
         onClick={keyword => {
-          history.push({ pathname: "/catalog", search: `tags=${keyword}` });
+          navigate({ pathname: "/catalog", search: `tags=${keyword}` });
           window.scrollTo({ top: 0, behavior: "smooth" });
         }}
       />
