@@ -1,138 +1,78 @@
 import {
-  FontWeights,
   IStackStyles,
-  ITextStyles,
   Stack,
-  Text,
   getTheme,
   IStackTokens,
-  StackItem,
-  FontSizes,
   IconButton,
   IButtonStyles,
 } from "@fluentui/react";
-import * as qs from "query-string";
 import { useLocalStorage } from "react-use";
 
-import { ILegendConfig } from "pages/Explore/types";
-import { LegendTypes } from "pages/Explore/enums";
 import { useExploreSelector } from "pages/Explore/state/hooks";
-import ColorMap from "./ColorMap";
-import ClassMap from "./ClassMap";
-import { hasClassmapValues } from "./helpers";
-import { IStacCollection } from "types/stac";
 import { controlStyle } from "../PanelControl";
+import Legend from "./Legend";
 
 export const LegendControl = () => {
-  const renderOpts = useExploreSelector(s => s.mosaic.renderOption);
-  const collection = useExploreSelector(s => s.mosaic.collection);
   const [isOpen, setIsOpen] = useLocalStorage("legend-control-open", true);
+  const { layers, layerOrder } = useExploreSelector(s => s.mosaic);
 
-  if (!renderOpts) return null;
+  // Generate legends for each layer, in order
+  const legends = layerOrder
+    .map(id => id && <Legend key={`legend-${id}`} layer={layers[id]} />)
+    .filter(Boolean);
 
-  const renderConfig = qs.parse(renderOpts.options || "");
-  const legendConfig = renderOpts.legend;
-  const legend = getLegendType(renderConfig, legendConfig, collection);
-  const renderDesc =
-    renderOpts.name && renderOpts?.name !== "Default" ? (
-      <Text block styles={subHeaderStyles}>
-        {renderOpts.name}
-      </Text>
-    ) : null;
-
+  // Keep the panel rendered even if it's closed to preserve the state of any command options
+  const openStyle = { display: isOpen ? "block" : "none" };
   const legendPanel = (
-    <Stack styles={panelStyles} tokens={stackTokens}>
-      <StackItem>
-        <Stack horizontal horizontalAlign="space-between">
-          <Text block styles={headerStyles}>
-            {collection?.title}
-          </Text>
-          <IconButton
-            title="Hide Legend"
-            iconProps={{ iconName: "ChevronDown" }}
-            styles={minimizeButtonStyles}
-            onClick={() => setIsOpen(!isOpen)}
-          />
-        </Stack>
-        {renderDesc}
-      </StackItem>
-      {legend}
+    <Stack
+      styles={panelStyles}
+      style={openStyle}
+      tokens={stackTokens}
+      data-cy={"explore-legend"}
+    >
+      {legends}
     </Stack>
   );
 
-  const buttonStyle = { right: 2, bottom: 32, ...controlStyle };
+  const legendTitle = isOpen ? "Hide Legend" : "Open Legend";
   const legendButton = (
-    <div style={buttonStyle}>
-      <IconButton
-        ariaLabel={legend ? "Open legend" : "No legend available"}
-        title={"Open legend"}
-        onClick={() => setIsOpen(true)}
-        disabled={!legend}
-        styles={legendButtonStyles}
-        className="azure-maps-control-button"
-        iconProps={{ iconName: "MapLegend" }}
-      />
+    <div style={buttonStyle} title={legendTitle}>
+      <div>
+        <IconButton
+          ariaLabel="Open legend"
+          className="azure-maps-control-button"
+          styles={legendButtonStyles}
+          iconProps={{ iconName: "MapLegend" }}
+          onClick={() => setIsOpen(!isOpen)}
+        />
+      </div>
     </div>
   );
 
-  return isOpen && legend ? legendPanel : legendButton;
-};
-
-const getLegendType = (
-  params: qs.ParsedQuery<string>,
-  legendConfig: ILegendConfig | undefined,
-  collection: IStacCollection | null
-) => {
-  // Legend configs are optional, but if they exist, use them to set the values
-  if (legendConfig) {
-    switch (legendConfig.type) {
-      case LegendTypes.classmap:
-        return <ClassMap params={params} collection={collection} />;
-      case LegendTypes.continuous:
-        return <ColorMap params={params} legendConfig={legendConfig} />;
-      case LegendTypes.none:
-        return null;
-      default:
-        throw new Error(`Unknown legend type: ${legendConfig.type}`);
-    }
-  }
-
-  if ("rescale" in params) {
-    return <ColorMap params={params} legendConfig={legendConfig} />;
-  }
-
-  if (hasClassmapValues(collection, params.assets)) {
-    return <ClassMap params={params} collection={collection} />;
-  }
+  const hasLegends = legends.length > 0;
+  return hasLegends ? (
+    <>
+      {legendPanel}
+      {legendButton}
+    </>
+  ) : null;
 };
 
 const theme = getTheme();
 const stackTokens: IStackTokens = {
-  childrenGap: theme.spacing.s1,
+  childrenGap: 2,
 };
 const panelStyles: IStackStyles = {
   root: {
     background: theme.semanticColors.bodyBackground,
-    padding: 10,
+    padding: "10px 0px 10px 0px",
     borderRadius: 2,
     position: "absolute",
     zIndex: 1,
-    bottom: 40,
+    bottom: 75,
     right: 10,
     boxShadow: "rgb(0 0 0 / 16%) 0 0 4px",
-    width: 300,
-  },
-};
-
-const headerStyles: ITextStyles = {
-  root: {
-    fontWeight: FontWeights.semibold,
-  },
-};
-
-const subHeaderStyles: ITextStyles = {
-  root: {
-    fontSize: FontSizes.smallPlus,
+    width: 370,
   },
 };
 
@@ -145,13 +85,4 @@ const legendButtonStyles: IButtonStyles = {
   },
 };
 
-const minimizeButtonStyles: IButtonStyles = {
-  root: {
-    width: 18,
-    height: 18,
-  },
-  icon: {
-    color: theme.semanticColors.bodyText,
-    fontSize: FontSizes.xSmall,
-  },
-};
+const buttonStyle = { right: 2, bottom: 32, ...controlStyle };
