@@ -1,14 +1,26 @@
 import axios from "axios";
+import { ILegendConfig } from "pages/Explore/types";
 import { QueryFunctionContext, useQuery } from "react-query";
-import { FileExtValues, IStacCollection } from "types/stac";
+import { QsParamType } from "types";
+import * as qs from "query-string";
+import {
+  ClassificationExtClasses,
+  FileExtValues,
+  IStacCollection,
+} from "types/stac";
 import { DATA_URL } from "utils/constants";
 
 export const rootColormapUrl = `${DATA_URL}/legend/colormap`;
 export const rootClassmapUrl = `${DATA_URL}/legend/classmap`;
+export const rootIntervalUrl = `${DATA_URL}/legend/interval`;
+
+// [[[min, max], [r,g,b,a]],...]
+export type IntervalMap = [[number, number], [number, number, number, number]][];
+export type ClassMap = Record<string, number[]>;
 
 export const hasClassmapValues = (
   collection: IStacCollection | null,
-  assets: string | string[] | null
+  assets: QsParamType
 ) => {
   // If the rendering options don't include an asset, it's not a categorical
   if (!assets) return false;
@@ -18,29 +30,64 @@ export const hasClassmapValues = (
 
   // Check the collection for item_asset for the selected asset, and see if it uses file:values
   // which defines the labels for the categorical values
-  return collection && "file:values" in collection.item_assets[assets];
+  const hasFileValues =
+    collection && "file:values" in collection.item_assets[assets];
+  const hasClassificationClasses =
+    collection && "classification:classes" in collection.item_assets[assets];
+
+  return hasFileValues || hasClassificationClasses;
 };
 
-export const useClassmap = (classmapName: string | null) => {
-  return useQuery(["classmap", classmapName], getClassmapByName, {
+export const fileValuesToClassificationClasses = (
+  fv: FileExtValues[] | undefined
+) => {
+  return fv?.map(f => {
+    return { value: f.values[0], description: f.summary };
+  });
+};
+
+export const useClassmap = (classmapName: string | null, config?: ILegendConfig) => {
+  return useQuery(["classmap", classmapName, config], getLegendMappingByName, {
     refetchOnWindowFocus: false,
     refetchOnMount: false,
     enabled: Boolean(classmapName),
   });
 };
 
-const getClassmapByName = async (
-  queryParam: QueryFunctionContext<["classmap", string | null]>
-): Promise<Record<string, number[]>> => {
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const [_, classmapName] = queryParam.queryKey;
-
-  return await (
-    await axios.get(`${rootClassmapUrl}/${classmapName}`)
-  ).data;
+export const getClassNameByValue = (
+  value: string,
+  classes: ClassificationExtClasses[]
+) => {
+  const matchingValue = classes.find(fv => fv.value === Number(value));
+  return matchingValue?.description;
 };
 
-export const getClassNameByValue = (value: string, fileValues: FileExtValues[]) => {
-  const matchingValue = fileValues.find(fv => fv.values.includes(Number(value)));
-  return matchingValue?.summary;
+export const useInterval = (intervalName: string | null, config?: ILegendConfig) => {
+  return useQuery(["interval", intervalName, config], getLegendMappingByName, {
+    refetchOnWindowFocus: false,
+    refetchOnMount: false,
+    enabled: Boolean(intervalName),
+  });
+};
+
+const getLegendMappingByName = async (
+  queryParam: QueryFunctionContext<
+    ["classmap" | "interval", string | null, ILegendConfig | undefined]
+  >
+): Promise<ClassMap> => {
+  const [legendType, classmapName, config] = queryParam.queryKey;
+
+  const qsConfig = config
+    ? "?" +
+      qs.stringify(
+        { trim_start: config.trimStart, trim_end: config.trimEnd },
+        { skipNull: true }
+      )
+    : "";
+
+  const rootUrl = legendType === "classmap" ? rootClassmapUrl : rootIntervalUrl;
+
+  return await (
+    await axios.get(`${rootUrl}/${classmapName}${qsConfig}`)
+  ).data;
 };
