@@ -1,8 +1,7 @@
 import { useEffect, useState } from "react";
-import { Checkbox, Separator, Slider, Stack } from "@fluentui/react";
+import { Checkbox, Stack } from "@fluentui/react";
 import * as atlas from "azure-maps-control";
 import PanelControl from "../PanelControl";
-import { mosaicLayerName } from "../../hooks/useMosaicLayer";
 import { DEFAULT_MAP_STYLE } from "pages/Explore/utils/constants";
 import CollectionBoundaryToggle from "./CollectionBoundaryToggle";
 interface MapsOptionsControlProps {
@@ -10,17 +9,8 @@ interface MapsOptionsControlProps {
 }
 
 const MapSettingsControl = ({ mapRef }: MapsOptionsControlProps) => {
-  const [opacity, setOpacity] = useState<number>(100);
   const [currentMapStyle, setCurrentMapStyle] = useState<string>(DEFAULT_MAP_STYLE);
   const [showLabels, setShowLabels] = useState<boolean>(true);
-
-  // Update opacity when slider changes
-  useEffect(() => {
-    const mosaicLayer = mapRef?.current?.layers.getLayerById(mosaicLayerName);
-    if (mosaicLayer) {
-      (mosaicLayer as atlas.layer.TileLayer).setOptions({ opacity: opacity / 100 });
-    }
-  }, [mapRef, opacity]);
 
   // Set label visibility when toggled or a new basemap style is selected
   useEffect(() => {
@@ -58,19 +48,8 @@ const MapSettingsControl = ({ mapRef }: MapsOptionsControlProps) => {
   const title = "Manage map settings";
 
   return (
-    <PanelControl label={title} iconName="FluentSettings" top={178}>
+    <PanelControl label={title} iconName="FluentSettingsFilled" top={178}>
       <Stack tokens={{ childrenGap: 10 }} styles={{ root: { padding: 4 } }}>
-        <Slider
-          min={0}
-          max={100}
-          step={1}
-          value={opacity}
-          onChange={value => setOpacity(value)}
-          label="Layer opacity"
-          valueFormat={(value: number) => `${value}%`}
-          styles={{ root: { width: "100%" } }}
-        />
-        <Separator />
         <Checkbox
           label="Show map labels"
           checked={showLabels}
