@@ -2,9 +2,8 @@ import { useCallback, useEffect, useState } from "react";
 import { useWindowSize } from "react-use";
 import atlas from "azure-maps-control";
 
-import { useExploreDispatch } from "pages/Explore/state/hooks";
+import { useExploreDispatch, useExploreSelector } from "pages/Explore/state/hooks";
 import { setCamera } from "pages/Explore/state/mapSlice";
-import { mosaicLayerName } from "./useMosaicLayer";
 import {
   collectionLineLayer,
   collectionLineLayerName,
@@ -12,11 +11,14 @@ import {
   itemLineLayerName,
   itemOutlineLayerName,
 } from "pages/Explore/utils/layers";
+import { makeLayerId } from "./useMosaicLayer";
 
 const useMapEvents = (mapRef: React.MutableRefObject<atlas.Map | null>) => {
   const dispatch = useExploreDispatch();
   const { height, width } = useWindowSize();
   const [areTilesLoading, setTilesLoading] = useState<boolean>(false);
+  const layerOrder = useExploreSelector(s => s.mosaic.layerOrder);
+  const bottomLayer = makeLayerId(layerOrder[layerOrder.length - 1]);
 
   useEffect(() => {
     // Some actions that resize the map that occur from browser chrome (show
@@ -39,37 +41,40 @@ const useMapEvents = (mapRef: React.MutableRefObject<atlas.Map | null>) => {
 
   // When the basemap style is changed, it changes the order of all loaded layers
   // which need to be manually reset.
-  const onStyleDataLoaded = useCallback((e: atlas.MapDataEvent) => {
-    if (e.dataType === "style") {
-      const layerMgr = e.map.layers;
+  const onStyleDataLoaded = useCallback(
+    (e: atlas.MapDataEvent) => {
+      if (e.dataType === "style") {
+        const layerMgr = e.map.layers;
 
-      if (layerMgr.getLayers()[0].getId() !== "base") {
-        const hasOutlineLayer = layerMgr.getLayerById(itemLineLayerName);
-        if (hasOutlineLayer) {
-          layerMgr.move(itemLineLayerName, "labels");
-          layerMgr.move(itemOutlineLayerName, itemLineLayerName);
-        }
-        const hasCollectionLayer = layerMgr.getLayerById(collectionLineLayerName);
-        if (hasCollectionLayer) {
-          layerMgr.move(collectionLineLayer, "labels");
-          layerMgr.move(collectionOutlineLayer, collectionLineLayer);
-        }
+        if (layerMgr.getLayers()[0].getId() !== "base") {
+          const hasOutlineLayer = layerMgr.getLayerById(itemLineLayerName);
+          if (hasOutlineLayer) {
+            layerMgr.move(itemLineLayerName, "labels");
+            layerMgr.move(itemOutlineLayerName, itemLineLayerName);
+          }
+          const hasCollectionLayer = layerMgr.getLayerById(collectionLineLayerName);
+          if (hasCollectionLayer) {
+            layerMgr.move(collectionLineLayer, "labels");
+            layerMgr.move(collectionOutlineLayer, collectionLineLayer);
+          }
 
-        const hasMosaicLayer = layerMgr.getLayerById(mosaicLayerName);
-        if (hasOutlineLayer && hasMosaicLayer) {
-          layerMgr.move(mosaicLayerName, itemLineLayerName);
-        }
+          const mosaicLayer = layerMgr.getLayerById(bottomLayer);
+          if (hasOutlineLayer && mosaicLayer) {
+            layerMgr.move(mosaicLayer, itemLineLayerName);
+          }
 
-        // To prevent runaway re-renders, move the base layer under the first
-        // layer if it isn't already. This likely means a custom layer hasn't been
-        // handled above.
-        const firstLayerId = layerMgr.getLayers()[0].getId();
-        if (firstLayerId !== "base") {
-          layerMgr.move("base", firstLayerId);
+          // To prevent runaway re-renders, move the base layer under the first
+          // layer if it isn't already. This likely means a custom layer hasn't been
+          // handled above.
+          const firstLayerId = layerMgr.getLayers()[0].getId();
+          if (firstLayerId !== "base") {
+            layerMgr.move("base", firstLayerId);
+          }
         }
       }
-    }
-  }, []);
+    },
+    [bottomLayer]
+  );
 
   // Loading indicator for mosaic tiles
   const onDataEvent = useCallback((e: atlas.MapDataEvent) => {
