@@ -2,25 +2,21 @@ import { IDropdownOption } from "@fluentui/react";
 import { useEffect } from "react";
 import { useCollectionMosaicInfo } from "../../../utils/hooks";
 import { useExploreDispatch, useExploreSelector } from "../../../state/hooks";
-import { setMosaicQuery } from "../../../state/mosaicSlice";
+import { selectCurrentMosaic, setMosaicQuery } from "../../../state/mosaicSlice";
 import StateSelector from "./StateSelector";
-import { useMosaicUrlState } from "./hooks/useUrlState";
 
 const MosaicPresetSelector = () => {
-  const { collection, query, isCustomQuery } = useExploreSelector(
-    state => state.mosaic
-  );
+  const { collection, query, isCustomQuery } =
+    useExploreSelector(selectCurrentMosaic);
   const dispatch = useExploreDispatch();
 
   const { isSuccess, data: mosaicInfo } = useCollectionMosaicInfo(collection?.id);
 
   useEffect(() => {
     if (mosaicInfo && query.name === null) {
-      dispatch<any>(setMosaicQuery(mosaicInfo.mosaics[0]));
+      dispatch(setMosaicQuery(mosaicInfo.mosaics[0]));
     }
   }, [dispatch, mosaicInfo, query.name]);
-
-  useMosaicUrlState(mosaicInfo?.mosaics);
 
   const mosaicOptions =
     isSuccess && mosaicInfo?.mosaics
