@@ -2,14 +2,15 @@ import { useEffect } from "react";
 import { IDropdownOption } from "@fluentui/react";
 
 import { useCollectionMosaicInfo } from "../../../utils/hooks";
-import { setRenderOption } from "../../../state/mosaicSlice";
+import { selectCurrentMosaic, setRenderOption } from "../../../state/mosaicSlice";
 import { useExploreDispatch, useExploreSelector } from "../../../state/hooks";
 import StateSelector from "./StateSelector";
-import { useRenderUrlState } from "./hooks/useUrlState";
+import useCqlPropertyMatcher from "./hooks/useCqlPropertyMatcher";
+import { every } from "lodash-es";
 
 const RenderOptionsSelector = () => {
   const dispatch = useExploreDispatch();
-  const { collection, renderOption } = useExploreSelector(state => state.mosaic);
+  const { collection, renderOption } = useExploreSelector(selectCurrentMosaic);
   const { data: mosaicInfo } = useCollectionMosaicInfo(collection?.id);
 
   // Set a default render option if none is selected
@@ -19,18 +20,36 @@ const RenderOptionsSelector = () => {
     }
   }, [dispatch, mosaicInfo, renderOption]);
 
-  useRenderUrlState(mosaicInfo?.renderOptions);
-
   const renderers = mosaicInfo?.renderOptions ?? [];
 
+  const matcher = useCqlPropertyMatcher();
   const options = renderers.map((renderer): IDropdownOption => {
-    return { key: renderer.name, text: renderer.name };
+    const enabled = every(
+      renderer.conditions?.map(condition =>
+        matcher(condition.property, condition.value)
+      )
+    );
+    return { key: renderer.name, text: renderer.name, disabled: !enabled };
   });
 
   const getOptionByName = (key: string | number) => {
     return mosaicInfo?.renderOptions?.find(mosaic => mosaic.name === key);
   };
 
+  // If the currently selected option is now disabled, move to the first
+  // non-disabled option
+  const enabledSelectedKey = options.find(
+    option => option.key === renderOption?.name
+  )?.disabled
+    ? options.find(option => !option.disabled)?.key
+    : renderOption?.name;
+
+  if (enabledSelectedKey !== renderOption?.name) {
+    if (enabledSelectedKey) {
+      const newOption = getOptionByName(enabledSelectedKey);
+      newOption && dispatch(setRenderOption(newOption));
+    }
+  }
   return (
     <StateSelector
       title="Select a rendering option"
