@@ -1,5 +1,4 @@
 import useStacFilter from "./useStacFilter";
 import { useCollectionMosaicInfo } from "./useCollectionMosaicInfo";
-import { useCollectionMapInfo } from "./useCollectionMapInfo";
 
-export { useStacFilter, useCollectionMosaicInfo, useCollectionMapInfo };
+export { useStacFilter, useCollectionMosaicInfo };
