@@ -1,5 +1,6 @@
 import { QueryFunctionContext, useQuery } from "react-query";
 import $RefParser, { JSONSchema } from "@apidevtools/json-schema-ref-parser";
+import { STAC_URL } from "utils/constants";
 
 export const useCollectionQueryables = (collectionId: string | undefined) => {
   return useQuery(["queryable", collectionId], getCollectionQueryables, {
@@ -13,11 +14,10 @@ export const useCollectionQueryables = (collectionId: string | undefined) => {
 const getCollectionQueryables = async (
   queryParam: QueryFunctionContext<["queryable", string | undefined]>
 ): Promise<JSONSchema> => {
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const [_, collectionId] = queryParam.queryKey;
+  const [, collectionId] = queryParam.queryKey;
 
   const schema = await $RefParser.dereference(
-    `/stac/${collectionId}/queryables.json`
+    `${STAC_URL}/collections/${collectionId}/queryables`
   );
 
   return schema;
