@@ -1,4 +1,5 @@
 import axios from "axios";
+// import { useSession } from "components/auth/hooks/SessionContext";
 import { QueryFunctionContext, useQuery, UseQueryResult } from "react-query";
 import { IStacFilter, IStacSearchResult } from "types/stac";
 import { STAC_URL } from "utils/constants";
@@ -6,8 +7,7 @@ import { STAC_URL } from "utils/constants";
 const getStacItems = async (
   queryParam: QueryFunctionContext<[string, IStacFilter | undefined]>
 ): Promise<IStacSearchResult> => {
-  // eslint-disable-next-line
-  const [_, search] = queryParam.queryKey;
+  const [, search] = queryParam.queryKey;
 
   if (typeof search === "undefined") {
     return Promise.reject();
@@ -21,6 +21,7 @@ const getStacItems = async (
 export const useStacSearch = (
   search: IStacFilter | undefined
 ): UseQueryResult<IStacSearchResult, Error> => {
+  // const { status } = useSession();
   return useQuery(["items", search], getStacItems, {
     keepPreviousData: true, // intended to not clear out search results when panning the map
     refetchOnWindowFocus: false,
