@@ -1,13 +1,8 @@
-import { render } from "@testing-library/react";
+import { render } from "testUtils";
 import { Home } from "./Home";
-import { MemoryRouter } from "react-router";
 
 test("smoke home render", () => {
-  const { getByText } = render(
-    <MemoryRouter>
-      <Home />
-    </MemoryRouter>
-  );
+  const { getByText } = render(<Home />);
 
   expect(
     getByText(/Planetary Computer for a Sustainable Future/i)
