@@ -64,9 +64,93 @@ const SearchIcon = (
   </svg>
 );
 
+const SettingsIconFilled = (
+  <svg viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path
+      d="M2.2672 6.15302C2.52892 5.34708 2.96005 4.60729 3.52891 3.98033C3.62511 3.87431 3.77603 3.83652 3.91115 3.88463L5.27062 4.36863C5.6392 4.49978 6.04476 4.30859 6.17647 3.9416C6.18942 3.90551 6.1994 3.86844 6.20629 3.8308L6.46529 2.41527C6.49104 2.27452 6.5996 2.1632 6.74013 2.13341C7.1513 2.04627 7.57289 2.00195 8 2.00195C8.42685 2.00195 8.84818 2.04621 9.25912 2.13325C9.39961 2.16301 9.50816 2.27427 9.53397 2.41496L9.7937 3.83074C9.86392 4.21413 10.233 4.46825 10.618 4.39833C10.6559 4.39146 10.6931 4.38153 10.7293 4.36865L12.0888 3.88463C12.224 3.83652 12.3749 3.87431 12.4711 3.98033C13.04 4.60729 13.4711 5.34708 13.7328 6.15302C13.777 6.28899 13.7344 6.43805 13.625 6.53053L12.5231 7.46205C12.2247 7.7143 12.1882 8.1596 12.4416 8.45667C12.4665 8.48589 12.4937 8.51303 12.5231 8.53782L13.625 9.46935C13.7344 9.56183 13.777 9.71089 13.7328 9.84685C13.4711 10.6528 13.04 11.3926 12.4711 12.0196C12.3749 12.1256 12.224 12.1634 12.0888 12.1153L10.7294 11.6312C10.3608 11.5001 9.95524 11.6913 9.82353 12.0583C9.81058 12.0944 9.8006 12.1314 9.79368 12.1692L9.53397 13.5849C9.50816 13.7256 9.39961 13.8369 9.25912 13.8666C8.84818 13.9537 8.42685 13.9979 8 13.9979C7.57289 13.9979 7.1513 13.9536 6.74013 13.8665C6.5996 13.8367 6.49104 13.7254 6.46529 13.5846L6.2063 12.1691C6.13608 11.7857 5.76701 11.5316 5.38196 11.6015C5.3441 11.6084 5.30687 11.6184 5.27068 11.6312L3.91115 12.1153C3.77603 12.1634 3.62511 12.1256 3.52891 12.0196C2.96005 11.3926 2.52892 10.6528 2.2672 9.84685C2.22305 9.71089 2.26562 9.56183 2.37502 9.46935L3.47693 8.53783C3.77528 8.28558 3.81177 7.84027 3.55843 7.5432C3.53352 7.51399 3.50626 7.48685 3.47694 7.46206L2.37502 6.53053C2.26562 6.43805 2.22305 6.28899 2.2672 6.15302ZM6.24988 7.99988C6.24988 8.96638 7.03338 9.74988 7.99988 9.74988C8.96638 9.74988 9.74988 8.96638 9.74988 7.99988C9.74988 7.03338 8.96638 6.24988 7.99988 6.24988C7.03338 6.24988 6.24988 7.03338 6.24988 7.99988Z"
+      fill="#83888d"
+    />
+  </svg>
+);
+
 const SettingsIcon = (
-  <svg viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path d="M1.91075 7.38266C2.28004 6.24053 2.88839 5.19213 3.69109 4.30364C3.82683 4.15339 4.03978 4.09984 4.23044 4.16802L6.14873 4.85392C6.6688 5.03977 7.24107 4.76883 7.42692 4.24875C7.4452 4.19762 7.45927 4.14507 7.469 4.09173L7.83446 2.08573C7.8708 1.88627 8.02398 1.7285 8.22227 1.6863C8.80246 1.5628 9.39734 1.5 10 1.5C10.6023 1.5 11.1968 1.56273 11.7767 1.68607C11.9749 1.72824 12.1281 1.88591 12.1645 2.08529L12.531 4.09165C12.6301 4.63497 13.1509 4.9951 13.6942 4.89601C13.7476 4.88627 13.8002 4.87219 13.8512 4.85395L15.7696 4.16802C15.9602 4.09984 16.1732 4.15339 16.3089 4.30364C17.1116 5.19213 17.72 6.24053 18.0893 7.38266C18.1516 7.57534 18.0915 7.78658 17.9371 7.91764L16.3823 9.23773C15.9613 9.5952 15.9098 10.2263 16.2673 10.6473C16.3024 10.6887 16.3409 10.7271 16.3823 10.7623L17.9371 12.0824C18.0915 12.2134 18.1516 12.4247 18.0893 12.6173C17.72 13.7595 17.1116 14.8079 16.3089 15.6964C16.1732 15.8466 15.9602 15.9002 15.7696 15.832L13.8513 15.1461C13.3312 14.9602 12.759 15.2312 12.5731 15.7512C12.5548 15.8024 12.5408 15.8549 12.531 15.9085L12.1645 17.9147C12.1281 18.1141 11.9749 18.2718 11.7767 18.3139C11.1968 18.4373 10.6023 18.5 10 18.5C9.39734 18.5 8.80246 18.4372 8.22227 18.3137C8.02398 18.2715 7.8708 18.1137 7.83446 17.9143L7.46902 15.9084C7.36993 15.365 6.84916 15.0049 6.30583 15.104C6.25241 15.1137 6.19987 15.1278 6.14881 15.1461L4.23044 15.832C4.03978 15.9002 3.82683 15.8466 3.69109 15.6964C2.88839 14.8079 2.28004 13.7595 1.91075 12.6173C1.84845 12.4247 1.90852 12.2134 2.06289 12.0824L3.61773 10.7623C4.03872 10.4048 4.09021 9.77373 3.73274 9.35274C3.69759 9.31135 3.65913 9.27288 3.61775 9.23775L2.06289 7.91764C1.90852 7.78658 1.84845 7.57534 1.91075 7.38266ZM8.00001 10C8.00001 11.1046 8.89544 12 10 12C11.1046 12 12 11.1046 12 10C12 8.89543 11.1046 8 10 8C8.89544 8 8.00001 8.89543 8.00001 10Z" />
+  <svg viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path
+      d="M7.99994 6C6.89537 6 5.99994 6.89543 5.99994 8C5.99994 9.10457 6.89537 10 7.99994 10C9.10451 10 9.99994 9.10457 9.99994 8C9.99994 6.89543 9.10451 6 7.99994 6ZM6.99994 8C6.99994 7.44772 7.44765 7 7.99994 7C8.55222 7 8.99994 7.44772 8.99994 8C8.99994 8.55228 8.55222 9 7.99994 9C7.44765 9 6.99994 8.55228 6.99994 8ZM10.618 4.39833C10.233 4.46825 9.86392 4.21413 9.7937 3.83074L9.53397 2.41496C9.50816 2.27427 9.39961 2.16301 9.25912 2.13325C8.84818 2.04621 8.42685 2.00195 8 2.00195C7.57289 2.00195 7.1513 2.04627 6.74013 2.13341C6.5996 2.1632 6.49104 2.27452 6.46529 2.41527L6.20629 3.8308C6.1994 3.86844 6.18942 3.90551 6.17647 3.9416C6.04476 4.30859 5.6392 4.49978 5.27062 4.36863L3.91115 3.88463C3.77603 3.83652 3.62511 3.87431 3.52891 3.98033C2.96005 4.60729 2.52892 5.34708 2.2672 6.15302C2.22305 6.28899 2.26562 6.43805 2.37502 6.53053L3.47694 7.46206C3.50626 7.48685 3.53352 7.51399 3.55843 7.5432C3.81177 7.84027 3.77528 8.28558 3.47693 8.53783L2.37502 9.46935C2.26562 9.56183 2.22305 9.71089 2.2672 9.84685C2.52892 10.6528 2.96005 11.3926 3.52891 12.0196C3.62511 12.1256 3.77603 12.1634 3.91115 12.1153L5.27068 11.6312C5.30687 11.6184 5.3441 11.6084 5.38196 11.6015C5.76701 11.5316 6.13608 11.7857 6.2063 12.1691L6.46529 13.5846C6.49104 13.7254 6.5996 13.8367 6.74013 13.8665C7.1513 13.9536 7.57289 13.9979 8 13.9979C8.42685 13.9979 8.84818 13.9537 9.25912 13.8666C9.39961 13.8369 9.50816 13.7256 9.53397 13.5849L9.79368 12.1692C9.8006 12.1314 9.81058 12.0944 9.82353 12.0583C9.95524 11.6913 10.3608 11.5001 10.7294 11.6312L12.0888 12.1153C12.224 12.1634 12.3749 12.1256 12.4711 12.0196C13.04 11.3926 13.4711 10.6528 13.7328 9.84685C13.777 9.71089 13.7344 9.56183 13.625 9.46935L12.5231 8.53782C12.4937 8.51303 12.4665 8.48589 12.4416 8.45667C12.1882 8.1596 12.2247 7.71429 12.5231 7.46205L13.625 6.53053C13.7344 6.43805 13.777 6.28899 13.7328 6.15302C13.4711 5.34708 13.04 4.60729 12.4711 3.98033C12.3749 3.87431 12.224 3.83652 12.0888 3.88463L10.7293 4.36865C10.6931 4.38152 10.6559 4.39146 10.618 4.39833ZM3.99863 4.97726L4.93522 5.3107C5.82017 5.62559 6.79872 5.16815 7.11769 4.2794C7.14903 4.19207 7.17324 4.1021 7.18996 4.01078L7.36738 3.04113C7.5757 3.01512 7.78684 3.00195 8 3.00195C8.213 3.00195 8.42397 3.0151 8.63214 3.04107L8.81011 4.01117C8.98053 4.9408 9.87266 5.55003 10.7967 5.38225C10.8877 5.36572 10.9775 5.34176 11.0647 5.31073L12.0014 4.97726C12.2564 5.31084 12.4684 5.67476 12.6319 6.06064L11.8774 6.6984C11.1566 7.30787 11.0675 8.38649 11.6807 9.10555C11.7408 9.17609 11.8067 9.24166 11.8775 9.3015L12.6319 9.93924C12.4684 10.3251 12.2564 10.689 12.0014 11.0226L11.0646 10.6891C10.1797 10.3742 9.20128 10.8317 8.88231 11.7205C8.85096 11.8078 8.82677 11.8978 8.81004 11.9891L8.63214 12.9588C8.42397 12.9848 8.213 12.9979 8 12.9979C7.78684 12.9979 7.5757 12.9848 7.36738 12.9587L7.18994 11.989C7.01965 11.0592 6.12743 10.4498 5.2033 10.6176C5.11227 10.6342 5.0225 10.6581 4.93528 10.6892L3.99863 11.0226C3.74357 10.689 3.53161 10.3251 3.36814 9.93924L4.12257 9.30148C4.84343 8.69201 4.93254 7.61339 4.31933 6.89433C4.25917 6.82378 4.19332 6.75822 4.12254 6.69838L3.36814 6.06064C3.53161 5.67476 3.74357 5.31084 3.99863 4.97726Z"
+      fill="#83888d"
+    />
+  </svg>
+);
+
+const View = (
+  <svg viewBox="0 0 18 18" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path
+      d="M2.25909 9.60214C2.94254 6.32689 5.79437 4 9.00002 4C12.2057 4 15.0574 6.32688 15.7409 9.60215C15.7974 9.87246 16.0622 10.0459 16.3325 9.98946C16.6029 9.93304 16.7763 9.66817 16.7199 9.39785C15.9425 5.67312 12.6934 3 9.00002 3C5.3066 3 2.05742 5.67311 1.28017 9.39786C1.22377 9.66818 1.39718 9.93305 1.6675 9.98946C1.93782 10.0459 2.20268 9.87246 2.25909 9.60214ZM9 6C7.067 6 5.5 7.567 5.5 9.5C5.5 11.433 7.067 13 9 13C10.933 13 12.5 11.433 12.5 9.5C12.5 7.567 10.933 6 9 6ZM6.5 9.5C6.5 8.11929 7.61929 7 9 7C10.3807 7 11.5 8.11929 11.5 9.5C11.5 10.8807 10.3807 12 9 12C7.61929 12 6.5 10.8807 6.5 9.5Z"
+      fill="#83888d"
+    />
+  </svg>
+);
+
+const UnView = (
+  <svg viewBox="0 0 18 18" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path
+      d="M0.853553 0.146447C0.658291 -0.0488155 0.341709 -0.0488155 0.146447 0.146447C-0.0488155 0.341709 -0.0488155 0.658291 0.146447 0.853553L3.64526 4.35237C1.97039 5.49178 0.723337 7.27383 0.280114 9.39787C0.223706 9.66819 0.397116 9.93305 0.667436 9.98946C0.937755 10.0459 1.20262 9.87246 1.25903 9.60214C1.66284 7.66698 2.82362 6.06289 4.3671 5.07421L5.94894 6.65605C5.06509 7.29133 4.48947 8.32844 4.48947 9.5C4.48947 11.433 6.05647 13 7.98947 13C9.16103 13 10.1981 12.4244 10.8334 11.5405L15.1464 15.8536C15.3417 16.0488 15.6583 16.0488 15.8536 15.8536C16.0488 15.6583 16.0488 15.3417 15.8536 15.1464L0.853553 0.146447ZM8.12378 6.00253L11.4869 9.3657C11.418 7.5395 9.94997 6.07143 8.12378 6.00253ZM5.53104 3.4098L6.3341 4.21286C6.87141 4.07353 7.43009 4 7.99995 4C11.2056 4 14.0574 6.32688 14.7409 9.60215C14.7973 9.87247 15.0622 10.0459 15.3325 9.98946C15.6028 9.93304 15.7762 9.66817 15.7198 9.39786C14.9425 5.67312 11.6934 3 7.99995 3C7.14478 3 6.31342 3.14331 5.53104 3.4098Z"
+      fill="#83888d"
+    />
+  </svg>
+);
+
+const Brightness = (
+  <svg
+    width="18"
+    height="18"
+    viewBox="0 0 18 18"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      d="M8.5 2.5C8.5 2.22386 8.27614 2 8 2C7.72386 2 7.5 2.22386 7.5 2.5V3.5C7.5 3.77614 7.72386 4 8 4C8.27614 4 8.5 3.77614 8.5 3.5V2.5ZM8 5C9.65685 5 11 6.34315 11 8C11 9.65685 9.65685 11 8 11C6.34315 11 5 9.65685 5 8C5 6.34315 6.34315 5 8 5ZM10 8C10 6.89543 9.10457 6 8 6V10C9.10457 10 10 9.10457 10 8ZM14 8C14 8.27614 13.7761 8.5 13.5 8.5H12.5C12.2239 8.5 12 8.27614 12 8C12 7.72386 12.2239 7.5 12.5 7.5H13.5C13.7761 7.5 14 7.72386 14 8ZM8.5 12.5C8.5 12.2239 8.27614 12 8 12C7.72386 12 7.5 12.2239 7.5 12.5V13.5C7.5 13.7761 7.72386 14 8 14C8.27614 14 8.5 13.7761 8.5 13.5V12.5ZM4 8C4 8.27614 3.77614 8.5 3.5 8.5H2.5C2.22386 8.5 2 8.27614 2 8C2 7.72386 2.22386 7.5 2.5 7.5H3.5C3.77614 7.5 4 7.72386 4 8ZM4.35355 3.64651C4.15829 3.45125 3.84171 3.45125 3.64645 3.64651C3.45118 3.84177 3.45118 4.15835 3.64645 4.35361L4.14645 4.85361C4.34171 5.04888 4.65829 5.04888 4.85355 4.85361C5.04882 4.65835 5.04882 4.34177 4.85355 4.14651L4.35355 3.64651ZM3.64645 12.3536C3.84171 12.5489 4.15829 12.5489 4.35355 12.3536L4.85355 11.8536C5.04882 11.6584 5.04882 11.3418 4.85355 11.1465C4.65829 10.9512 4.34171 10.9512 4.14645 11.1465L3.64645 11.6465C3.45118 11.8418 3.45118 12.1584 3.64645 12.3536ZM11.6464 3.64651C11.8417 3.45125 12.1583 3.45125 12.3536 3.64651C12.5488 3.84177 12.5488 4.15835 12.3536 4.35361L11.8536 4.85361C11.6583 5.04888 11.3417 5.04888 11.1464 4.85361C10.9512 4.65835 10.9512 4.34177 11.1464 4.14651L11.6464 3.64651ZM11.6464 12.3536L11.1464 11.8536C10.9512 11.6584 10.9512 11.3418 11.1464 11.1465C11.3417 10.9512 11.6583 10.9512 11.8536 11.1465L12.3536 11.6465C12.5488 11.8418 12.5488 12.1584 12.3536 12.3536C12.1583 12.5489 11.8417 12.5489 11.6464 12.3536Z"
+      fill="#83888d"
+    />
+  </svg>
+);
+
+const Warning = (
+  <svg width="16" height="16" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path
+      d="M2.56367 12.2571L7.56524 3.25712C7.73988 2.94286 8.16874 2.91667 8.38575 3.17856L8.43933 3.25712L13.4409 12.2571C13.6119 12.5648 13.4197 12.9374 13.0892 12.993L13.0039 13H3.00072C2.64878 13 2.41637 12.651 2.52837 12.3352L2.56367 12.2571L7.56524 3.25712L2.56367 12.2571ZM8.00229 4.52932L3.85061 12H12.154L8.00229 4.52932ZM8.00229 9.87694C8.34641 9.87694 8.62537 10.1559 8.62537 10.5C8.62537 10.8441 8.34641 11.1231 8.00229 11.1231C7.65816 11.1231 7.3792 10.8441 7.3792 10.5C7.3792 10.1559 7.65816 9.87694 8.00229 9.87694ZM8.00191 7.00257C8.24737 7.00275 8.45138 7.17979 8.49354 7.41307L8.50152 7.50295L8.50076 8.50038C8.50055 8.77652 8.27652 9.00021 8.00038 9C7.75492 8.99981 7.5509 8.82278 7.50875 8.5895L7.50076 8.49962L7.50152 7.50218C7.50173 7.22604 7.72576 7.00235 8.00191 7.00257Z"
+      fill="#83888d"
+    />
+  </svg>
+);
+
+const FluentPinFilled = (
+  <svg
+    width="18"
+    height="18"
+    viewBox="0 0 18 18"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      d="M16.4347 5.78709L12.2122 1.56455C11.0508 0.403188 9.09299 0.720901 8.35842 2.18992L6.22746 6.45194C6.15073 6.60541 6.01695 6.72272 5.85477 6.77874L2.20928 8.03807C1.44567 8.30188 1.22175 9.27401 1.79301 9.8453L4.50944 12.5617L0.62507 16.4461L0.625 17.3742H1.55315L5.43752 13.4898L8.15394 16.2063C8.72522 16.7775 9.69735 16.5536 9.96116 15.79L11.2205 12.1444C11.2765 11.9823 11.3939 11.8485 11.5473 11.7718L15.8093 9.64081C17.2783 8.90625 17.5961 6.94844 16.4347 5.78709Z"
+      fill="#196CC9"
+    />
+  </svg>
+);
+
+const FluentPinOutline = (
+  <svg
+    width="16"
+    height="16"
+    viewBox="0 0 18 18"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      d="M12.237 1.48578L16.5141 5.76293C17.6905 6.9393 17.3687 8.92242 15.8807 9.66648L11.5635 11.825C11.4082 11.9027 11.2893 12.0383 11.2326 12.2025L9.9569 15.8952C9.68968 16.6686 8.70498 16.8955 8.1263 16.3169L5.37475 13.5652L1.44015 17.4999H0.5L0.500071 16.5598L4.43468 12.6252L1.68312 9.87362C1.10447 9.29494 1.33128 8.31024 2.10477 8.04302L5.7974 6.7674C5.96168 6.71066 6.09719 6.59183 6.17491 6.43638L8.33344 2.11924C9.0775 0.63122 11.0606 0.309398 12.237 1.48578ZM15.5741 6.70301L11.2969 2.42586C10.7622 1.89114 9.86074 2.03742 9.52261 2.7138L7.36407 7.03093C7.13088 7.49728 6.72433 7.85379 6.2315 8.02396L2.87366 9.18397L8.81595 15.1263L9.97596 11.7685C10.1462 11.2756 10.5027 10.8691 10.969 10.6359L15.2861 8.47731C15.9625 8.13918 16.1088 7.23772 15.5741 6.70301Z"
+      fill="rgb(0, 120, 212)"
+    />
   </svg>
 );
 
@@ -79,7 +163,14 @@ export const registerCustomIcons = () => {
       FluentFilter: FluentFilterIcon,
       MapLegend: LegendIcon,
       FluentSearch: SearchIcon,
+      FluentSettingsFilled: SettingsIconFilled,
       FluentSettings: SettingsIcon,
+      FluentView: View,
+      FluentUnview: UnView,
+      FluentBrightness: Brightness,
+      FluentWarning: Warning,
+      FluentPinFilled,
+      FluentPinOutline,
     },
   });
 };
